Add comprehensive tests for reaction handling and router functionality
- Introduced tests for the reaction handling logic in `tests/test_handlers_reactions.py`, covering various scenarios including bot user ID caching, reaction event processing, and authorization checks for deleting bot messages.
- Added tests for the router functionality in `tests/test_router.py`, validating the request routing, app caching, and worker processing, ensuring proper handling of Slack events and credentials management.
